Don't call DeleteHostSubnetRules on a replayed Deleted event
Fixes bug 1547599 (https://bugzilla.redhat.com/show_bug.cgi?id=1547599)

+++ pkg/sdn/plugin/subnets.go
@@ -307,6 +307,9 @@ func (node *OsdnNode) watchSubnets() {
 			node.AddHostSubnetRules(hs)
 			subnets[string(hs.UID)] = hs
 		case cache.Deleted:
+			if _, exists := subnets[string(hs.UID)]; !exists {
+				return nil
+			}
 			delete(subnets, string(hs.UID))
 			node.DeleteHostSubnetRules(hs)
 		}
